+++ src/app/views/shared/dynamic-library-examples/fhi-angular-examples/fhi-angular-highcharts-example/fhi-angular-highcharts-example.component.ts
@@ -79,6 +79,7 @@ export class FhiAngularHighchartsExampleComponent implements OnInit {
             creditsText: 'Nøkkeltall for folkehelse',
             disclaimer: 'Disse dataene kan inneholde feil.',
             lastUpdated: '06.06.2023',
+            mapTypeId: 'mapFylker',
             openSource: false
           };
           this.dataIsLoading = false;